Feature/replace expired time type to local data time

+++ api/src/main/java/com/workshop/vote/domain/factories/NewTopicFactory.java
@@ -5,15 +5,17 @@ import com.workshop.vote.infra.crossCutting.patterns.factory.BaseFactoryWithPara
 import org.javatuples.Quartet;
 import org.springframework.stereotype.Component;
 
+import java.time.ZonedDateTime;
+
 @Component
-public class NewTopicFactory extends BaseFactoryWithParams<Quartet<String, Double, String, String>, NewTopic> {
+public class NewTopicFactory extends BaseFactoryWithParams<Quartet<String, Long, String, String>, NewTopic> {
 
     @Override
-    public NewTopic create(Quartet<String, Double, String, String> params) {
+    public NewTopic create(Quartet<String, Long, String, String> params) {
         var entity = new NewTopic();
         entity.setNewInfoRegister(
             params.getValue0(), // Topic name
-            params.getValue1(), // secondDuration
+            params.getValue1(), // expirationTime
             params.getValue2(), // createdBy
             params.getValue3() // lastPlatformSource
         );
